requirements.txt
langgraph
langchain-core
langchain-deepseek
langchain_community
langchain_openai
langchain-tavily
python-dotenv
langsmith
pydantic
matplotlib
seaborn
pandas
pymysql
scikit-learn
akshare
django 
djangorestframework 
pytest 
pytest-django